more unit testing

+++ commotion-service-manager.h
@@ -102,7 +102,7 @@ void resolve_callback(
   const char *domain,
   const char *host_name,
   const AvahiAddress *address,
-  uint16_t port,
+  long port,
   AvahiStringList *txt,
   AvahiLookupResultFlags flags,
   void* userdata);
